requirements.txt
fastapi
uvicorn
python-dotenv
google-cloud-translate
